renamed FNotepadLANG.java to FNotepad.java
